feat(pulse) added supporting type pulsemeta (#10)
PulseMeta ends up being used internally in the larger ecosystem quite a bit. It's worth noting here that there's an attempt at separation on what's in Pulse vs PulseMeta that is essentially, "the things that are the identity of a pulse" vs "the things that help systems track and use pulses". Also! Being borne out of a larger system PulseMeta didn't really need any kind of extendability. That gave the benefit of strong type safety. I would love to be able to open it up more to end use cases but every way I tried to do that in experimenting before this public release came with tradeoffs I was unhappy with. So for now, tags are about all it gets. I'd love to see union types hit Swift someday though.