feat(pulse): added base pulse type (#11)
This commit doesn't include the fluent builders or the factory methods. Just trying to get thins in with tiny pieces to help me make sure I'm sorting it right as i pull it out of the other codebase. Yay Pulse though! I also moved PulseMeta and PulseSource to be in a subdirectory to be a little clearer in organization.