feat(pulse): static resppond and instance like (#19)
Additions to the public API for creating pulses. Respond is a common case of creating a base pulse that echoes another, it's really a convenience method. The instance `like` method is useful for batch pulses or other cases where the majority of (corrected with chaining) or all metadata is useful to copy.