ci(github): first pass at github actions to test (#24)
* ci(github): first pass at github actions to test
I was planning on getting CI setup before merging into main anyway but wasn't hella worried about it. Then I did the thing with the package being broken on release and decided to make this priority. I've not used actions with swifft before though so this is an attempt based on past experience and the GitHub swift actions guide: https://docs.github.com/en/actions/use-cases-and-examples/building-and-testing/building-and-testing-swift

* fix(ci): gotta use the correct tag syntax babe

* deps(ci:setup-swift): current version is broken for linux
I'll check back on this later and see if we can get other versions working. A bit out of scope rn. Anyway see: [setup-swift](https://github.com/swift-actions/setup-swift/issues/694)