ci(github): support macos 14 and 15 (#25)
* ci(github): can we run on recent macos versions
Mostly a for fun test. I am curious. :D

* ci(github): separate build vs test
This is fun! But running the full test suite for each OS is probably a bit unnecessary. Also the last run had issues with macos 13 (intel) so that may be a thing. Anyway, let's try this.

* ci(macos): lower compatibility to 14
Ok so I think we can settle on this. Since the runner had issues on tests in macos 13 and I have no real way of working through that I will just roll with 14 for now. It's helpful that this exists because it means I can lower the required minimum version of the package itself. Love that. ^^