build(qlty): remove qlty (#31)
I really like CodeClimate for insights on code complexity and areas to DRY up and such. Unfortunately qlty and the current support for Swift isn't there. The intent was to keep in in place and it would automagically start working when they added support but... That's kind of silly. PURGE IT!!! \o/