feat(channel): breaking - remove key based release
**BREAKING CHANGE**

The original implementation of Channel (not oss project) didn't use key based auth for release. I had anxiety around the possibility of a channel getting passed and a sender calling it.

The reality is this has very little real world benefit and is a massive pain to keep around in primitives that build on channel - which is the primary expected usage pattern of Sable.

Channels aren't intended to be passed around or have multiple senders without being very sure of what someone is doing. So at the risk of making some poor design choices easier, I'm relaxing this requirement to make all other code also easier.