feat(stream): add stream to sable
Streams!!! Streams are built on Channels using chain-of-trust and provide bidirectional lifecycle handling of both sides of the stream. They're a core component for other, incoming, Sable features and are expected to be the primary driver of most systems. Channels are still useful independently when lighter weight solutions are needed, of course. <3